// generated/src/app/api-models/med-defn-alt-srch-params.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MedDefnAltSrchParams {
  criteriaSrchPriority: number;
  altSrchCriteria: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  medorDisplayColumn: string;
}